requirements.txt
streamlit
pandas
openpyxl
groq
python-dotenv
tqdm

